backend/requirements.txt
# DB and ORM
aerich==0.7.2
tortoise-orm==0.20.0
aiosqlite==0.17.0

# Taskiq and redis
redis==5.0.1
aioredis==2.0.1
taskiq==0.10.2
taskiq-redis==0.5.3

# API
fastapi==0.104.0
fastapi-admin==1.0.4
fastapi-jwt==0.2.0
uvicorn==0.23.2
PyJWT==2.8.0
httpx==0.25.1
itsdangerous==2.1.2

# Other
aiosmtplib==3.0.1
Authlib==1.2.1
email-validator==2.1.0.post1
python-dotenv==1.0.0


